test: PoolingNestedStorageTest
