+++ module/data/src/main/kotlin/io/github/siyual_park/data/repository/mongo/MongoRepositoryBuilder.kt
@@ -3,7 +3,7 @@ package io.github.siyual_park.data.repository.mongo
 import com.google.common.cache.CacheBuilder
 import io.github.siyual_park.data.cache.InMemoryStorage
 import io.github.siyual_park.data.cache.LoadingPool
-import io.github.siyual_park.data.cache.NestedStorage
+import io.github.siyual_park.data.cache.PoolingNestedStorage
 import io.github.siyual_park.data.cache.TransactionalStorage
 import io.github.siyual_park.data.repository.Extractor
 import io.github.siyual_park.event.EventPublisher
@@ -44,7 +44,7 @@ class MongoRepositoryBuilder<T : Any, ID : Any>(
         return if (cacheBuilder != null) {
             val idExtractor = createIdExtractor(current)
             val storage = TransactionalStorage(
-                NestedStorage(LoadingPool { InMemoryStorage(cacheBuilder, idExtractor) }, idExtractor)
+                PoolingNestedStorage(LoadingPool { InMemoryStorage(cacheBuilder, idExtractor) }, idExtractor)
             )
 
             CachedMongoRepository(current, storage, idExtractor)
