+++ module/data/src/main/kotlin/io/github/siyual_park/data/repository/r2dbc/R2DBCRepositoryBuilder.kt
@@ -4,8 +4,8 @@ import com.google.common.cache.CacheBuilder
 import io.github.siyual_park.data.cache.InMemoryQueryStorage
 import io.github.siyual_park.data.cache.InMemoryStorage
 import io.github.siyual_park.data.cache.LoadingPool
-import io.github.siyual_park.data.cache.NestedQueryStorage
-import io.github.siyual_park.data.cache.NestedStorage
+import io.github.siyual_park.data.cache.PoolingNestedQueryStorage
+import io.github.siyual_park.data.cache.PoolingNestedStorage
 import io.github.siyual_park.data.cache.TransactionalQueryStorage
 import io.github.siyual_park.data.cache.TransactionalStorage
 import io.github.siyual_park.data.repository.Extractor
@@ -47,7 +47,7 @@ class R2DBCRepositoryBuilder<T : Any, ID : Any>(
             if (cacheBuilder != null) {
                 val idExtractor = createIdExtractor(it)
                 val storage = TransactionalStorage(
-                    NestedStorage(LoadingPool { InMemoryStorage(cacheBuilder, idExtractor) }, idExtractor)
+                    PoolingNestedStorage(LoadingPool { InMemoryStorage(cacheBuilder, idExtractor) }, idExtractor)
                 )
 
                 CachedR2DBCRepository(it, storage, idExtractor)
@@ -58,7 +58,7 @@ class R2DBCRepositoryBuilder<T : Any, ID : Any>(
             val queryCacheBuilder = queryCacheBuilder
             if (queryCacheBuilder != null) {
                 val storage = TransactionalQueryStorage<T>(
-                    NestedQueryStorage(LoadingPool { InMemoryQueryStorage(queryCacheBuilder) })
+                    PoolingNestedQueryStorage(LoadingPool { InMemoryQueryStorage(queryCacheBuilder) })
                 )
 
                 CachedQueryR2DBCRepository(it, storage)
