+++ module/data/src/test/kotlin/io/github/siyual_park/data/cache/PoolingNestedStorageTest.kt
@@ -1,0 +1,25 @@
+package io.github.siyual_park.data.cache
+
+import com.google.common.cache.CacheBuilder
+import io.github.siyual_park.data.entity.Person
+import io.github.siyual_park.data.repository.Extractor
+import io.github.siyual_park.ulid.ULID
+
+private fun create(): PoolingNestedStorage<Person, ULID> {
+    val idExtractor = object : Extractor<Person, ULID> {
+        override fun getKey(entity: Person): ULID {
+            return entity.id
+        }
+    }
+    return PoolingNestedStorage(
+        LoadingPool {
+            InMemoryStorage(
+                { CacheBuilder.newBuilder() },
+                idExtractor
+            )
+        },
+        idExtractor
+    )
+}
+
+class PoolingNestedStorageTest : NestedStorageTestHelper(create())
