+++ module/data/src/test/kotlin/io/github/siyual_park/data/cache/TransactionalQueryStorageTest.kt
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test
 
 class TransactionalQueryStorageTest : CoroutineTestHelper() {
     private val storage = TransactionalQueryStorage<String>(
-        NestedQueryStorage(
+        PoolingNestedQueryStorage(
             LoadingPool { InMemoryQueryStorage { CacheBuilder.newBuilder() } }
         )
     )
